+++ platform/viewer/src/googleCloud/utils/getServers.js
@@ -29,4 +29,31 @@ const getServers = (data, name) => {
   ];
 };
 
-export { getServers };
+const isValidServer = server => {
+  return (
+    server &&
+    !!server.dataset &&
+    !!server.dicomStore &&
+    !!server.location &&
+    !!server.project
+  );
+};
+
+const isEqualServer = (server = {}, toCompare = {}) => {
+  const serverLength = Object.keys(server).length;
+  const toCompareLength = Object.keys(toCompare).length;
+
+  if (!serverLength || !toCompareLength) {
+    return false;
+  }
+
+  return (
+    server.dataset === toCompare.dataset &&
+    server.dataset === toCompare.dataset &&
+    server.dicomStore === toCompare.dicomStore &&
+    server.location === toCompare.location &&
+    server.project === toCompare.project
+  );
+};
+
+export { getServers, isValidServer, isEqualServer };